+++ src/js/requests/nearestRequest.js
@@ -28,7 +28,10 @@ module.exports = class nearestRequest extends Request {
   constructor(resource, coordinates) {
 
     // Constructeur parent
-    super("nearest", resource, "nearestRequest");
+    super("nearest", "nearestRequest");
+
+    // Ressource concernée
+    this._resource = resource;
 
     // Coordonnées du point fourni
     this._coordinates = coordinates;
@@ -36,6 +39,29 @@ module.exports = class nearestRequest extends Request {
     // Nombre de points attendus en retour 
     this._number = 1;
 
+  }
+
+  /**
+  *
+  * @function
+  * @name get resource
+  * @description Récupérer la ressource de la requête
+  *
+  */
+   get resource () {
+    return this._resource;
+  }
+
+  /**
+  *
+  * @function
+  * @name set resource
+  * @description Attribuer la ressource de la requête
+  * @param {string} res - Id de la ressource
+  *
+  */
+  set resource (res) {
+    this._resource = res;
   }
 
   /**