+++ src/js/service/main.js
@@ -80,6 +80,12 @@ async function startService() {
 
           LOGGER.info("Les sources connectables ont été connectées");
 
+          // Instanciation de la fonction permettant de recevoir des messages via IPC
+          if (process.argv[3] === "child") {
+            LOGGER.info("Ce service est un child d'un administrateur. Instanciation de la fonction permettant de recevoir les messages");
+            service.initIPC();
+          }
+
           // On démarre les serveurs associé à ce service
           if (!service.startServers()) {
             pm.shutdown(5);