+++ src/js/service/serviceInsider.js
@@ -143,4 +143,21 @@ module.exports = class ServiceInsider extends ServiceAdministered {
 
   }
 
+  /**
+   *
+   * @function
+   * @name computeRequest
+   * @description Fonction pour utiliser pour envoyer une requête à un service selon le mode adaptée à la classe fille. Elle doit être ré-écrite dans chaque classe fille.
+   * @param {object} request - Instance fille de la classe Request 
+   * 
+   */
+  computeRequest(request) {
+
+    LOGGER.info("computeRequest...");
+
+    // L'instance de Service est accessible, il suffit de faire appel à la fonction qui traite les requêtes
+    return this._serviceInstance.computeAdminRequest(request);
+
+  }
+
 }