+++ src/js/sources/source.js
@@ -42,10 +42,12 @@ module.exports = class Source {
     // État de la connexion de la source
     this._connected = false;
 
-    // État de la source (même si connectée, elle peut être disfonctionnelle)
-    // Peut être : "green" si la dernière requête a fonctionnée, "orange" si la source est connectée mais injoignable, "red" à l'initialisation ou si plus gros problème
+    // État de la source (même si connectée, elle peut être dysfonctionnelle)
+    // "green" si la dernière requête a fonctionnée, 
+    // "red" si la donnée n'est plus accessible, 
+    // "init" à sa création (utile pour une requête /health de l'administrateur)
     // Ajouter la gestion de ce paramètre dans chaque classe fille
-    this._state = "red";
+    this._state = "init";
 
   }