+++ changelog.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2.1.0
+
+CHANGED:
+  - La documentation de l'API d'administration a été grandement enrichie. 
+  - La route /health a une réponse plus complète et est vraiment codée pour prendre en compte l'état de chaque service et chaque source disponibles. 
+
 ## 2.0.0
 
 ADDED: